- Provided "android.permission.MANAGE_EXTERNAL_STORAGE" permission access to the user and removed from the library
