+++ app/src/main/AndroidManifest.xml
@@ -2,14 +2,17 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.lassi.app">
 
+    <!--        This permission require only for document picker feature and if Android device SDK is >= 30-->
+    <uses-permission android:name="android.permission.MANAGE_EXTERNAL_STORAGE" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:preserveLegacyExternalStorage="true"
         android:requestLegacyExternalStorage="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:preserveLegacyExternalStorage="true"
         android:theme="@style/AppTheme">
 
         <activity
